+++ app/src/main/java/com/control_delivery/finanzas_delivery/domain/repository/TimeBasedExpenseRepository.kt
@@ -13,7 +13,7 @@ interface TimeBasedExpenseRepository {
      * Get a flow of all the today expenses.
      * @param today The date to check.
      */
-    fun getDailyAmountExpenses(today: LocalDate = LocalDate.now()): Flow<Double>
+    fun getDailyAmountExpenses(today: LocalDate = LocalDate.now()): Flow<Long>
 
     /**
      * Get a flow of a specific expense by its ID.
